Restore ServerPoller class
