+++ Source/RimConnection/RimConnection.csproj
@@ -114,7 +114,7 @@
     <Compile Include="Settings\ResetCommandOptionsModal.cs" />
     <Compile Include="Settings\CommandOptionSettings.cs" />
     <Compile Include="Managers\AlertManager.cs" />
-    <Compile Include="DonationAlertsPoller.cs" />
+    <Compile Include="ServerPoller.cs" />
     <Compile Include="API\DonationAlertsAPI.cs" />
     <Compile Include="Managers\PawnCreationManager.cs" />
     <Compile Include="RimConnection.cs" />
